Fixed Bugs in controllers
