+++ app/Http/Controllers/Admin/HomeController.php
@@ -5,7 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use Illuminate\View\View;
 
-class AdminHomeController extends Controller
+class HomeController extends Controller
 {
     public function index(): View
     {
