+++ resources/views/admin/challenge.blade.php
