+++ resources/views/admin/product.blade.php
@@ -1,7 +1,7 @@
 @extends('layouts.admin')
-@section('headerEntity', 'Users')
-@section('nameTable', 'Users')
-@section('buttonEntity', 'user')
+@section('headerEntity', 'Products')
+@section('nameTable', 'Products')
+@section('buttonEntity', 'product')
 @section('content')
     @if (session('success'))
         <div class="bg-green-100 border border-green-400 text-green-700 px-4 py-3 rounded relative" role="alert">
@@ -10,63 +10,79 @@
         </div>
     @endif
     <div
-        class="p-4 bg-white block sm:flex items-center justify-between border-b border-gray-200 lg:mt-1.5 dark:bg-gray-800 dark:border-gray-700">
+        class='p-4 bg-white block sm:flex items-center justify-between border-b border-gray-200 lg:mt-1.5 dark:bg-gray-800 dark:border-gray-700'>
         <!--Table Header-->
         @include('partials.admin.adminTableHeader')
     </div>
-    <div class="flex flex-col">
-        <div class="overflow-x-auto">
-            <div class="inline-block min-w-full align-middle">
-                <div class="overflow-hidden shadow">
-                    <table class="min-w-full divide-y divide-gray-200 table-fixed dark:divide-gray-600">
-                        <thead class="bg-gray-100 dark:bg-gray-700">
+    <div class='flex flex-col'>
+        <div class='overflow-x-auto'>
+            <div class='inline-block min-w-full align-middle'>
+                <div class='overflow-hidden shadow'>
+                    <table class='min-w-full divide-y divide-gray-200 table-fixed dark:divide-gray-600'>
+                        <thead class='bg-gray-100 dark:bg-gray-700'>
                             <!--Columns -->
                             <tr>
-                                <th scope="col"
-                                    class="p-4 text-xs font-medium text-left text-gray-500 uppercase dark:text-gray-400">
-                                    NickName
+                                <th scope='col'
+                                    class='p-4 text-xs font-medium text-left text-gray-500 uppercase dark:text-gray-400'>
+                                    Product Name
                                 </th>
                                 <th scope="col"
                                     class="p-4 text-xs font-medium text-left text-gray-500 uppercase dark:text-gray-400">
-                                    Email
+                                    Brand
                                 </th>
                                 <th scope="col"
                                     class="p-4 text-xs font-medium text-left text-gray-500 uppercase dark:text-gray-400">
-                                    Balance
+                                    Keywords
                                 </th>
                                 <th scope="col"
                                     class="p-4 text-xs font-medium text-left text-gray-500 uppercase dark:text-gray-400">
-                                    Role
+                                    Stock
+                                </th>
+                                <th scope='col'
+                                    class='p-4 text-xs font-medium text-left text-gray-500 uppercase dark:text-gray-400'>
+                                    Price
                                 </th>
                                 <th scope="col"
                                     class="p-4 text-xs font-medium text-left text-gray-500 uppercase dark:text-gray-400">
+                                    Description
+                                </th>
+                                <th scope='col'
+                                    class='p-4 text-xs font-medium text-left text-gray-500 uppercase dark:text-gray-400'>
                                     Actions
                                 </th>
                             </tr>
                         </thead>
                         <tbody class="bg-white divide-y divide-gray-200 dark:bg-gray-800 dark:divide-gray-700">
-                            @foreach ($users as $user)
+                            @foreach ($products as $product)
                                 <tr class="hover:bg-gray-100 dark:hover:bg-gray-700">
                                     <!-- Object Values -->
                                     <td class="p-4 text-sm font-normal text-gray-500 whitespace-nowrap dark:text-gray-400">
                                         <div class="text-base font-semibold text-gray-900 dark:text-white">
-                                            {{ $user->getName() }}</div>
+                                            {{ $product->getName() }}</div>
+                                        <div class="text-sm font-normal text-gray-500 dark:text-gray-400">
+                                            {{ $product->category->getName() }}</div>
                                     </td>
                                     <td class="p-4 text-base font-medium text-gray-900 whitespace-nowrap dark:text-white">
-                                        {{ $user->getEmail() }}</td>
+                                        {{ $product->getBrand() }}</td>
+                                    <td
+                                        class="max-w-sm p-4 overflow-hidden text-base font-normal text-gray-500 truncate xl:max-w-xs dark:text-gray-400">
+                                        {{ implode(', ', $product->getKeywords()) }} </td>
                                     <td class="p-4 text-base font-medium text-gray-900 whitespace-nowrap dark:text-white">
-                                        {{ $user->getBalance() }}
+                                        {{ $product->getStock() }}
                                     </td>
                                     <td class="p-4 text-base font-medium text-gray-900 whitespace-nowrap dark:text-white">
-                                        {{ $user->getRole() }}
+                                        {{ $product->getPrice() }}
+                                    </td>
+                                    <td class="p-4 text-base font-medium text-gray-900 whitespace-nowrap dark:text-white">
+                                        {{ $product->getDescription() }}
                                     </td>
 
                                     <!-- Actions -->
                                     <td class="p-4 space-x-2 whitespace-nowrap">
-                                        <button type="button" id="updateUserButton"
-                                            data-drawer-target="drawer-update-user-{{ $user->getId() }}"
-                                            data-drawer-show="drawer-update-user-{{ $user->getId() }}"
-                                            aria-controls="drawer-update-user-{{ $user->getId() }}"
+                                        <button type="button" id="updateProductButton"
+                                            data-drawer-target="drawer-update-product-{{ $product->getId() }}"
+                                            data-drawer-show="drawer-update-product-{{ $product->getId() }}"
+                                            aria-controls="drawer-update-product-{{ $product->getId() }}"
                                             data-drawer-placement="right"
                                             class="inline-flex items-center px-3 py-2 text-sm font-medium text-center text-white rounded-lg bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:ring-blue-300 dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800">
                                             <svg class="w-4 h-4 mr-2" fill="currentColor" viewBox="0 0 20 20"
@@ -81,10 +97,10 @@
                                             Update
                                         </button>
 
-                                        <button type="button" id="deleteUserButton"
-                                            data-drawer-target="drawer-delete-user-{{ $user->getId() }}"
-                                            data-drawer-show="drawer-delete-user-{{ $user->getId() }}"
-                                            aria-controls="drawer-delete-user-{{ $user->getId() }}"
+                                        <button type="button" id="deleteProductButton"
+                                            data-drawer-target="drawer-delete-product-{{ $product->getId() }}"
+                                            data-drawer-show="drawer-delete-product-{{ $product->getId() }}"
+                                            aria-controls="drawer-delete-product-{{ $product->getId() }}"
                                             data-drawer-placement="right"
                                             class="inline-flex items-center px-3 py-2 text-sm font-medium text-center text-white bg-red-700 rounded-lg hover:bg-red-800 focus:ring-4 focus:ring-red-300 dark:focus:ring-red-900">
                                             <svg class="w-4 h-4 mr-2" fill="currentColor" viewBox="0 0 20 20"
@@ -99,15 +115,17 @@
                                 </tr>
 
                                 <!-- Edit Product Drawer -->
-                                <div id="drawer-update-user-{{ $user->getId() }}"
+                                <div id="drawer-update-product-{{ $product->getId() }}"
                                     class="fixed top-0 right-0 z-40 w-full h-screen max-w-xs p-4 overflow-y-auto transition-transform translate-x-full bg-white dark:bg-gray-800"
-                                    tabindex="-1" aria-labelledby="drawer-label-{{ $user->getId() }}" aria-hidden="true">
-                                    <h5 id="drawer-label-{{ $user->getId() }}"
+                                    tabindex="-1" aria-labelledby="drawer-label-{{ $product->getId() }}"
+                                    aria-hidden="true">
+                                    <h5 id="drawer-label-{{ $product->getId() }}"
                                         class="inline-flex items-center mb-6 text-sm font-semibold text-gray-500 uppercase dark:text-gray-400">
                                         Update
-                                        User</h5>
-                                    <button type="button" data-drawer-dismiss="drawer-update-user-{{ $user->getId() }}"
-                                        aria-controls="drawer-update-user-{{ $user->getId() }}"
+                                        Product</h5>
+                                    <button type="button"
+                                        data-drawer-dismiss="drawer-update-product-{{ $product->getId() }}"
+                                        aria-controls="drawer-update-product-{{ $product->getId() }}"
                                         class="text-gray-400 bg-transparent hover:bg-gray-200 hover:text-gray-900 rounded-lg text-sm p-1.5 absolute top-2.5 right-2.5 inline-flex items-center dark:hover:bg-gray-600 dark:hover:text-white">
                                         <svg aria-hidden="true" class="w-5 h-5" fill="currentColor" viewBox="0 0 20 20"
                                             xmlns="http://www.w3.org/2000/svg">
@@ -117,54 +135,88 @@
                                         </svg>
                                         <span class="sr-only">Close menu</span>
                                     </button>
-                                    <form action="{{ route('admin.users.update', ['id' => $user->getId()]) }}"
-                                        id="update-form-{{ $user->getId() }}" method='POST'>
+                                    <form action="{{ route('admin.products.update', ['id' => $product->getId()]) }}"
+                                        id="update-form-{{ $product->getId() }}" method='POST'>
                                         @csrf
                                         @method('PUT')
                                         <div class="space-y-4">
                                             <div>
-                                                <label for="name-{{ $user->getId() }}"
+                                                <label for="name-{{ $product->getId() }}"
                                                     class="block mb-2 text-sm font-medium text-gray-900 dark:text-white">Name</label>
-                                                <input type="text" name="name" id="name-{{ $user->getId() }}"
+                                                <input type="text" name="name" id="name-{{ $product->getId() }}"
                                                     class="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-primary-600 focus:border-primary-600 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-primary-500 dark:focus:border-primary-500"
-                                                    value="{{ $user->getName() }}" placeholder="Type product name"
+                                                    value="{{ $product->getName() }}" placeholder="Type product name"
                                                     required="">
                                             </div>
                                             <div>
-                                                <label for="role-{{ $user->getId() }}"
-                                                    class="block mb-2 text-sm font-medium text-gray-900 dark:text-white">Role</label>
-                                                <select id="role-{{ $user->getId() }}" name="role"
+                                                <label for="category_id"
+                                                    class="block mb-2 text-sm font-medium text-gray-900 dark:text-white">Category</label>
+                                                <select id="category_id"
                                                     class="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-primary-500 focus:border-primary-500 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-primary-500 dark:focus:border-primary-500">
-                                                    <option value="admin"
-                                                        {{ $user->getRole() === 'admin' ? 'selected' : '' }}>Admin</option>
-                                                    <option value="client"
-                                                        {{ $user->getRole() === 'client' ? 'selected' : '' }}>Client
-                                                    </option>
+                                                    @foreach ($categories as $category)
+                                                        <option value="{{ $category->getId() }}"
+                                                            {{ $category->getId() == $product->category->getId() ? 'selected' : '' }}>
+                                                            {{ $category->getName() }}</option>
+                                                    @endforeach
                                                 </select>
                                             </div>
+                                            <div>
+                                                <label for="price-{{ $product->getId() }}"
+                                                    class="block mb-2 text-sm font-medium text-gray-900 dark:text-white">Price</label>
+                                                <input type="text" name="price" id="price-{{ $product->getId() }}"
+                                                    class="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-primary-600 focus:border-primary-600 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-primary-500 dark:focus:border-primary-500"
+                                                    value="{{ $product->getPrice() }}" placeholder="Price product "
+                                                    required="">
+                                            </div>
+                                            <div>
+                                                <label for="description-{{ $product->getId() }}"
+                                                    class="block mb-2 text-sm font-medium text-gray-900 dark:text-white">Description</label>
+                                                <input type="text" name="description"
+                                                    id="description-{{ $product->getId() }}"
+                                                    class="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-primary-600 focus:border-primary-600 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-primary-500 dark:focus:border-primary-500"
+                                                    value="{{ $product->getDescription() }}"
+                                                    placeholder="Type product name" required="">
+                                            </div>
+                                            <div>
+                                                <label for="stock-{{ $product->getId() }}"
+                                                    class="block mb-2 text-sm font-medium text-gray-900 dark:text-white">Stock</label>
+                                                <input type="text" name="stock" id="stock-{{ $product->getId() }}"
+                                                    class="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-primary-600 focus:border-primary-600 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-primary-500 dark:focus:border-primary-500"
+                                                    value="{{ $product->getStock() }}" placeholder="Type product name"
+                                                    required="">
+                                            </div>
+                                            <div>
+                                                <label for="brand-{{ $product->getId() }}"
+                                                    class="block mb-2 text-sm font-medium text-gray-900 dark:text-white">Brand</label>
+                                                <input type="text" name="brand" id="brand-{{ $product->getId() }}"
+                                                    class="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-primary-600 focus:border-primary-600 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-primary-500 dark:focus:border-primary-500"
+                                                    value="{{ $product->getBrand() }}" placeholder="Type product name"
+                                                    required="">
+                                            </div>
+                                            <div>
+                                                <label for="keywords-{{ $product->getId() }}"
+                                                    class="block mb-2 text-sm font-medium text-gray-900 dark:text-white">Keywords</label>
+                                                <input type="text" name="keywords"
+                                                    id="keywords-{{ $product->getId() }}"
+                                                    class="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-primary-600 focus:border-primary-600 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-primary-500 dark:focus:border-primary-500"
+                                                    value="{{ implode(', ', $product->getKeywords()) }}"
+                                                    placeholder="Type product name" required="">
+                                            </div>
+                                            <div>
+                                                <label for="image-{{ $product->getId() }}"
+                                                    class="block mb-2 text-sm font-medium text-gray-900 dark:text-white">Image</label>
+                                                <input type="text" name="image" id="image-{{ $product->getId() }}"
+                                                    class="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-primary-600 focus:border-primary-600 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-primary-500 dark:focus:border-primary-500"
+                                                    value="{{ $product->getImage() }}" placeholder="Type product name"
+                                                    required="">
+                                            </div>
 
-                                            <div>
-                                                <label for="email-{{ $user->getId() }}"
-                                                    class="block mb-2 text-sm font-medium text-gray-900 dark:text-white">Email</label>
-                                                <input type="email" name="email" id="email-{{ $user->getId() }}"
-                                                    class="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-primary-600 focus:border-primary-600 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-primary-500 dark:focus:border-primary-500"
-                                                    value="{{ $user->getEmail() }}" placeholder="Price product "
-                                                    required="">
-                                            </div>
-                                            <div>
-                                                <label for="balance-{{ $user->getId() }}"
-                                                    class="block mb-2 text-sm font-medium text-gray-900 dark:text-white">Balance</label>
-                                                <input type="numeric" name="balance" id="balance-{{ $user->getId() }}"
-                                                    class="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-primary-600 focus:border-primary-600 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-primary-500 dark:focus:border-primary-500"
-                                                    value="{{ $user->getBalance() }}" placeholder="Type product name"
-                                                    required="">
-                                            </div>
                                         </div>
                                         <div class="flex justify-center mt-10">
-                                            <button type="submit" id="updateUserButton"
-                                                data-drawer-target="drawer-update-user-{{ $user->getId() }}"
-                                                data-drawer-show="drawer-update-user-{{ $user->getId() }}"
-                                                aria-controls="drawer-update-user-{{ $user->getId() }}"
+                                            <button type="submit" id="updateProductButton"
+                                                data-drawer-target="drawer-update-product-default"
+                                                data-drawer-show="drawer-update-product-default"
+                                                aria-controls="drawer-update-product-default"
                                                 data-drawer-placement="right"
                                                 class="inline-flex items-center px-3 py-2 text-sm font-medium text-center text-white rounded-lg bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:ring-blue-300 dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800">
                                                 <svg class="w-4 h-4 mr-2" fill="currentColor" viewBox="0 0 20 20"
@@ -181,17 +233,17 @@
                                         </div>
                                     </form>
                                 </div>
-
                                 <!-- Delete Product Drawer -->
-                                <div id="drawer-delete-user-{{ $user->getId() }}"
+                                <div id="drawer-delete-product-{{ $product->getId() }}""
                                     class="fixed top-0 right-0 z-40 w-full h-screen max-w-xs p-4 overflow-y-auto transition-transform translate-x-full bg-white dark:bg-gray-800"
                                     tabindex="-1" aria-labelledby="drawer-label" aria-hidden="true">
                                     <h5 id="drawer-label"
                                         class="inline-flex items-center text-sm font-semibold text-gray-500 uppercase dark:text-gray-400">
                                         Delete item
                                     </h5>
-                                    <button type="button" data-drawer-dismiss="drawer-delete-user-{{ $user->getId() }}"
-                                        aria-controls="drawer-delete-user-{{ $user->getId() }}"
+                                    <button type="button"
+                                        data-drawer-dismiss="drawer-delete-product-{{ $product->getId() }}""
+                                        aria-controls="drawer-delete-product-{{ $product->getId() }}""
                                         class="text-gray-400 bg-transparent hover:bg-gray-200 hover:text-gray-900 rounded-lg text-sm p-1.5 absolute top-2.5 right-2.5 inline-flex items-center dark:hover:bg-gray-600 dark:hover:text-white">
                                         <svg aria-hidden="true" class="w-5 h-5" fill="currentColor" viewBox="0 0 20 20"
                                             xmlns="http://www.w3.org/2000/svg">
@@ -207,9 +259,9 @@
                                             d="M12 8v4m0 4h.01M21 12a9 9 0 11-18 0 9 9 0 0118 0z"></path>
                                     </svg>
                                     <h3 class="mb-6 text-lg text-gray-500 dark:text-gray-400">Are you sure you want to
-                                        delete this User?</h3>
-                                    <form id="delete-form-{{ $user->getId() }}"
-                                        action="{{ route('admin.users.destroy', ['id' => $user->getId()]) }}) }}"
+                                        delete this product?</h3>
+                                    <form id="delete-form-{{ $product->getId() }}"
+                                        action="{{ route('admin.products.destroy', ['id' => $product->getId()]) }}"
                                         method='POST'>
                                         @csrf
                                         @method('DELETE')
@@ -220,7 +272,7 @@
                                     </form>
                                     <a href="#"
                                         class="text-gray-900 bg-white hover:bg-gray-100 focus:ring-4 focus:ring-primary-300 border border-gray-200 font-medium inline-flex items-center rounded-lg text-sm px-3 py-2.5 text-center dark:bg-gray-800 dark:text-gray-400 dark:border-gray-600 dark:hover:text-white dark:hover:bg-gray-700 dark:focus:ring-gray-700"
-                                        data-drawer-hide="drawer-delete-user-{{ $user->getId() }}">
+                                        data-drawer-hide="drawer-delete-product-default">
                                         No, cancel
                                     </a>
                                 </div>
@@ -230,31 +282,34 @@
                 </div>
             </div>
         </div>
+        <!-- Pagination -->
+        <div class="flex justify-center">
+            @include('partials.pagination', ['paginator' => $products])
+        </div>
     </div>
 
-    <!-- Pagination -->
-    @include('partials.pagination', ['paginator' => $users])
+
 
 
     <!-- Add Product Drawer -->
-    <div id="drawer-create-product-default"
-        class="fixed top-0 right-0 z-40 w-full h-screen max-w-xs p-4 overflow-y-auto transition-transform translate-x-full bg-white dark:bg-gray-800"
-        tabindex="-1" aria-labelledby="drawer-label" aria-hidden="true">
-        <h5 id="drawer-label"
-            class="inline-flex items-center mb-6 text-sm font-semibold text-gray-500 uppercase dark:text-gray-400">New
-            User</h5>
-        <button type="button" data-drawer-dismiss="drawer-create-product-default"
-            aria-controls="drawer-create-product-default"
-            class="text-gray-400 bg-transparent hover:bg-gray-200 hover:text-gray-900 rounded-lg text-sm p-1.5 absolute top-2.5 right-2.5 inline-flex items-center dark:hover:bg-gray-600 dark:hover:text-white">
-            <svg aria-hidden="true" class="w-5 h-5" fill="currentColor" viewBox="0 0 20 20"
-                xmlns="http://www.w3.org/2000/svg">
-                <path fill-rule="evenodd"
-                    d="M4.293 4.293a1 1 0 011.414 0L10 8.586l4.293-4.293a1 1 0 111.414 1.414L11.414 10l4.293 4.293a1 1 0 01-1.414 1.414L10 11.414l-4.293 4.293a1 1 0 01-1.414-1.414L8.586 10 4.293 5.707a1 1 0 010-1.414z"
-                    clip-rule="evenodd"></path>
+    <div id='drawer-create-product-default'
+        class='fixed top-0 right-0 z-40 w-full h-screen max-w-xs p-4 overflow-y-auto transition-transform translate-x-full bg-white dark:bg-gray-800'
+        tabindex='-1' aria-labelledby='drawer-label' aria-hidden='true'>
+        <h5 id='drawer-label'
+            class='inline-flex items-center mb-6 text-sm font-semibold text-gray-500 uppercase dark:text-gray-400'>New
+            Product</h5>
+        <button type='button' data-drawer-dismiss='drawer-create-product-default'
+            aria-controls='drawer-create-product-default'
+            class='text-gray-400 bg-transparent hover:bg-gray-200 hover:text-gray-900 rounded-lg text-sm p-1.5 absolute top-2.5 right-2.5 inline-flex items-center dark:hover:bg-gray-600 dark:hover:text-white'>
+            <svg aria-hidden='true' class='w-5 h-5' fill='currentColor' viewBox='0 0 20 20'
+                xmlns='http://www.w3.org/2000/svg'>
+                <path fill-rule='evenodd'
+                    d='M4.293 4.293a1 1 0 011.414 0L10 8.586l4.293-4.293a1 1 0 111.414 1.414L11.414 10l4.293 4.293a1 1 0 01-1.414 1.414L10 11.414l-4.293 4.293a1 1 0 01-1.414-1.414L8.586 10 4.293 5.707a1 1 0 010-1.414z'
+                    clip-rule='evenodd'></path>
             </svg>
-            <span class="sr-only">Close menu</span>
+            <span class='sr-only'>Close menu</span>
         </button>
-        <form action="{{ route('admin.users.store') }}" method='POST'>
+        <form action="{{ route('admin.products.store') }}" method='POST'>
             @csrf
             <div class="space-y-4">
                 <div>
@@ -265,42 +320,64 @@
                         placeholder="Type product name" required="">
                 </div>
                 <div>
-                    <label for="email"
-                        class="block mb-2 text-sm font-medium text-gray-900 dark:text-white">Email</label>
-                    <input type="email" name="email" id="email"
+                    <label for="description"
+                        class="block mb-2 text-sm font-medium text-gray-900 dark:text-white">Description</label>
+                    <input type="text" name="description" id="description"
                         class="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-primary-600 focus:border-primary-600 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-primary-500 dark:focus:border-primary-500"
                         placeholder="Type product name" required="">
                 </div>
                 <div>
-                    <label for="balance"
-                        class="block mb-2 text-sm font-medium text-gray-900 dark:text-white">Balance</label>
-                    <input type="numeric" name="balance" id="balance"
+                    <label for="brand"
+                        class="block mb-2 text-sm font-medium text-gray-900 dark:text-white">Brand</label>
+                    <input type="text" name="brand" id="brand"
                         class="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-primary-600 focus:border-primary-600 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-primary-500 dark:focus:border-primary-500"
                         placeholder="Type product name" required="">
                 </div>
                 <div>
-                    <label for="password"
-                        class="block mb-2 text-sm font-medium text-gray-900 dark:text-white">Password</label>
-                    <input type="password" name="password" id="password"
+                    <label for="price"
+                        class="block mb-2 text-sm font-medium text-gray-900 dark:text-white">Price</label>
+                    <input type="numeric" name="price" id="price"
                         class="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-primary-600 focus:border-primary-600 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-primary-500 dark:focus:border-primary-500"
                         placeholder="Type product name" required="">
                 </div>
                 <div>
-                    <label for="role"
-                        class="block mb-2 text-sm font-medium text-gray-900 dark:text-white">Role</label>
-                    <select name="role" id="role"
+                    <label for="stock"
+                        class="block mb-2 text-sm font-medium text-gray-900 dark:text-white">Stock</label>
+                    <input type="numeric" name="stock" id="stock"
+                        class="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-primary-600 focus:border-primary-600 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-primary-500 dark:focus:border-primary-500"
+                        placeholder="Type product name" required="">
+                </div>
+                <div>
+                    <label for="keywords"
+                        class="block mb-2 text-sm font-medium text-gray-900 dark:text-white">Keywords</label>
+                    <input type="text" name="keywords" id="keywords"
+                        class="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-primary-600 focus:border-primary-600 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-primary-500 dark:focus:border-primary-500"
+                        placeholder="Type product name" required="">
+                </div>
+                <div>
+                    <label for="image"
+                        class="block mb-2 text-sm font-medium text-gray-900 dark:text-white">Image</label>
+                    <input type="text" name="image" id="image"
+                        class="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-primary-600 focus:border-primary-600 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-primary-500 dark:focus:border-primary-500"
+                        placeholder="Type product name" required="">
+                </div>
+                <div>
+                    <label for="category_id"
+                        class="block mb-2 text-sm font-medium text-gray-900 dark:text-white">Category</label>
+                    <select name="category_id" id="category_id"
                         class="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-primary-500 focus:border-primary-500 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-primary-500 dark:focus:border-primary-500">
-                        <option>admin</option>
-                        <option>client</option>
+                        @foreach ($categories as $category)
+                            <option value="{{ $category->getId() }}">{{ $category->getName() }}</option>
+                        @endforeach
                     </select>
                 </div>
                 <div class="flex justify-center mt-10">
-                    <button id="createUserButton"
+                    <button id="createProductButton"
                         class="text-black bg-white hover:bg-gray-200 focus:ring-4 focus:ring-primary-300 font-medium rounded-lg text-sm px-5 py-2.5 border border-gray-300 focus:outline-none"
                         type="submit" data-drawer-target="drawer-create-product-default"
                         data-drawer-show="drawer-create-product-default" aria-controls="drawer-create-product-default"
                         data-drawer-placement="right">
-                        Add new user
+                        Add new product
                     </button>
                 </div>
             </div>
