--- resources/views/category/create.blade.php
@@ -1,15 +1,0 @@
-<!-- resources/views/categories/create.blade.php -->
-
-@extends('layouts.app')
-
-@section('content')
-    <h1>Create Category</h1>
-    <form action="{{ route('categories.store') }}" method="POST">
-        @csrf
-        <div class="form-group">
-            <label for="name">Category Name:</label>
-            <input type="text" name="name" id="name" class="form-control" required>
-        </div>
-        <button type="submit" class="btn btn-primary">Create Category</button>
-    </form>
-@endsection
