--- resources/views/category/edit.blade.php
@@ -1,14 +1,0 @@
-<!-- resources/views/categories/edit.blade.php -->
-
-@extends('layouts.app')
-
-@section('content')
-    <h1>Edit Category</h1>
-    <form action="{{ route('categories.update', $category->getId()) }}" method="POST">
-        @csrf
-        @method('PUT')
-        <label for="name">Name:</label>
-        <input type="text" id="name" name="name" value="{{ $category->getName() }}">
-        <button type="submit">Update</button>
-    </form>
-@endsection
