+++ resources/views/partials/admin/adminSidebar.blade.php
@@ -32,7 +32,7 @@
                         </a>
                     </li>
                     <li>
-                        <a href='#'
+                        <a href='{{ route('admin.challenges.index') }}'
                             class="flex items-center w-full p-2 text-base text-gray-900 transition duration-75 rounded-lg group hover:bg-gray-100 dark:text-gray-200 dark:hover:bg-gray-700"
                             aria-controls="dropdown-crud" data-collapse-toggle="dropdown-crud">
                             <svg class="flex-shrink-0 w-6 h-6 text-gray-500 transition duration-75 group-hover:text-gray-900 dark:text-gray-400 dark:group-hover:text-white"
